pyproject.toml
[project]
name = "bingomaker"
version = "0.1.0"
description = "A multiplayer bingo game maker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35.50",
    "flask>=3.0.3",
    "pillow>=11.0.0",
]

[tool.uv]
dev-dependencies = [
    "mongomock>=4.2.0.post1",
    "pytest>=8.3.3",
    "ruff>=0.7.1",
    "gunicorn>=23.0.0",
    "setproctitle>=1.3.3",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.format]
quote-style = "double"
